wip: refactor laws loading
